Using organization feature of latest keycloak version
Keycloak has a "new" feature called organization for multi-tenancy. I'm changing the project to use this feature and assign the customer entity to the organization that the authenticated user belongs. Due to these changes, I'm moving the common code used in the authentication flow to another lib called auth-lib that can be re-used and better cohesion.

+++ customer-svc/src/main/java/com/company/secureapispring/customer/services/CustomerService.java
@@ -1,5 +1,7 @@
 package com.company.secureapispring.customer.services;
 
+import com.company.secureapispring.auth.entities.Organization;
+import com.company.secureapispring.auth.services.AuthService;
 import com.company.secureapispring.customer.entities.Customer;
 import com.company.secureapispring.customer.repositories.CustomerRepository;
 import jakarta.persistence.EntityNotFoundException;
@@ -15,32 +17,39 @@ import java.util.List;
 @Transactional
 @RequiredArgsConstructor
 public class CustomerService {
+    private final AuthService authService;
     private final CustomerRepository customerRepository;
 
     public Customer create(Customer input) {
         Customer customer = new Customer();
         BeanUtils.copyProperties(input, customer);
+        authService.setOrganization(customer);
         return customerRepository.save(customer);
     }
 
     public Customer update(Customer input) {
         Customer customer = customerRepository.findById(input.getId()).orElseThrow(EntityNotFoundException::new);
+        authService.validateOwnership(customer);
         BeanUtils.copyProperties(input, customer);
         return customerRepository.save(customer);
     }
 
     public Customer delete(Long id) {
         Customer customer = customerRepository.findByIdWithStateProvince(id).orElseThrow(EntityNotFoundException::new);
+        authService.validateOwnership(customer);
         customerRepository.delete(customer);
         return customer;
     }
 
     public Customer get(Long id) {
-        return customerRepository.findByIdWithStateProvince(id)
+        Customer customer = customerRepository.findByIdWithStateProvince(id)
                 .orElseThrow(EntityNotFoundException::new);
+        authService.validateOwnership(customer);
+        return customer;
     }
 
     public List<Customer> listAll() {
-        return customerRepository.findAll(Sort.by("lastName", "firstName"));
+        Organization organization = authService.getAuthenticatedOrganization();
+        return customerRepository.findAllByOrganizationId(organization.getId(), Sort.by("lastName", "firstName"));
     }
 }
